Changed Shared proj to a class lib so everone uses the same IPlugin.cs
and updated all dependencies and usages of class lib
and testet it
