+++ SmartHub/SmartHub.Api.sln
@@ -25,11 +25,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{1F8BFE47
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{5982C7DF-4F10-4BBF-84C4-9B5C897ADBE7}"
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "SmartHub.BasePlugin", "SmartHub.BasePlugin\SmartHub.BasePlugin.shproj", "{80AD1FA2-346A-4E2C-9678-1CFC9B4B45A8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SmartHub.BasePlugin", "SmartHub.BasePlugin\SmartHub.BasePlugin.csproj", "{B1FC6183-41B0-439E-BA46-53B1DE88DF17}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		SmartHub.BasePlugin\SmartHub.BasePlugin.projitems*{80ad1fa2-346a-4e2c-9678-1cfc9b4b45a8}*SharedItemsImports = 13
+		SmartHub.BasePlugin\SmartHub.BasePlugin.projitems*{9e9fa8f4-9c11-4fce-8455-e9e96ef532ee}*SharedItemsImports = 5
 		SmartHub.BasePlugin\SmartHub.BasePlugin.projitems*{e666aab0-95c1-4638-a952-97109bbf2ab4}*SharedItemsImports = 5
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,8 +54,10 @@ Global
 		{E666AAB0-95C1-4638-A952-97109BBF2AB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E666AAB0-95C1-4638-A952-97109BBF2AB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E666AAB0-95C1-4638-A952-97109BBF2AB4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{80AD1FA2-346A-4E2C-9678-1CFC9B4B45A8}.Debug|Any CPU.ActiveCfg = Build|Any CPU
-		{80AD1FA2-346A-4E2C-9678-1CFC9B4B45A8}.Debug|Any CPU.Build.0 = Build|Any CPU
+		{B1FC6183-41B0-439E-BA46-53B1DE88DF17}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1FC6183-41B0-439E-BA46-53B1DE88DF17}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1FC6183-41B0-439E-BA46-53B1DE88DF17}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1FC6183-41B0-439E-BA46-53B1DE88DF17}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -64,7 +67,7 @@ Global
 		{106B3ACC-8E85-4961-9CC1-F6931322CACA} = {A4808DAE-2F25-4874-9D4A-778A02F10C05}
 		{1097B98A-DECF-43AB-84B7-5BD315CC417B} = {A4808DAE-2F25-4874-9D4A-778A02F10C05}
 		{E666AAB0-95C1-4638-A952-97109BBF2AB4} = {A4808DAE-2F25-4874-9D4A-778A02F10C05}
-		{80AD1FA2-346A-4E2C-9678-1CFC9B4B45A8} = {5982C7DF-4F10-4BBF-84C4-9B5C897ADBE7}
+		{B1FC6183-41B0-439E-BA46-53B1DE88DF17} = {5982C7DF-4F10-4BBF-84C4-9B5C897ADBE7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3B05BA04-E698-45A7-86F3-CD424A0D81C2}
