+++ SmartHub/SmartHub.Api/SmartHub.Api.csproj
@@ -72,8 +72,6 @@
     <ProjectReference Include="..\SmartHub.Infrastructure\SmartHub.Infrastructure.csproj" />
   </ItemGroup>
 
-  <Import Project="..\SmartHub.BasePlugin\SmartHub.BasePlugin.projitems" Label="Shared" />
-
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
     <!-- Ensure Node.js is installed -->
     <Exec Command="node --version" ContinueOnError="true">
