--- SmartHub/SmartHub.BasePlugin/Interfaces/DeviceType/IDoor.cs
@@ -1,7 +1,0 @@
-﻿namespace SmartHub.BasePlugin.Interfaces.DeviceType
-{
-	public interface IDoor : IBuild<IDoor>
-	{
-
-	}
-}
