+++ SmartHub/SmartHub.BasePlugin/Interfaces/DeviceTypes/IDoor.cs
@@ -1,0 +1,7 @@
+﻿namespace SmartHub.BasePlugin.Interfaces.DeviceTypes
+{
+    public interface IDoor : IBuild<IDoor>
+    {
+
+    }
+}
