+++ SmartHub/SmartHub.BasePlugin/Interfaces/IBuild.cs
@@ -1,8 +1,8 @@
 ﻿namespace SmartHub.BasePlugin.Interfaces
 {
-	public interface IBuild<T> : IPlugin where T : IBuild<T>
-	{
-		T Instantiate();
-		string Build();
-	}
+    public interface IBuild<T> : IPlugin where T : IBuild<T>
+    {
+        T Instantiate();
+        string Build();
+    }
 }
