+++ SmartHub/SmartHub.BasePlugin/Interfaces/IMqttSupport.cs
@@ -1,7 +1,7 @@
 ﻿namespace SmartHub.BasePlugin.Interfaces
 {
-	public interface IMqttSupport
-	{
-		string ExecuteMqtt();
-	}
+    public interface IMqttSupport
+    {
+        string ExecuteMqtt();
+    }
 }
